Disable collision-protect in favor of protect-owned.
Portage has two flags for doing collision protection: collision-protect
and protect-owned. The protect-owned feature is enabled by default and
is quite useful: it checks to make sure that we don't have multiple
packages that own the same file. The collision-protect feature is more
strict, and less useful: it fails if it finds a conflicting file, even
if that file was created by an earlier ebuild that failed to install.

BUG=none
TEST=emerge-x86-generic --info | grep FEATURES

Change-Id: Ia987c1fdb55b6c5b4bbc52638f0ec55864f6c6e0

Review URL: http://codereview.chromium.org/3421006