build_packages: extend kernel-headers workaround
This change extends the hacks in build_packages that were added
to handle the transition of the kernel-headers from the
kernel ebuild back into the kernel-headers ebuild.

Developers would still get a collision if they were
working on the kernel ebuild. This CL fixes that.

BUG=6734
TEST=Verified before and after cases if working on kernel.

Change-Id: Ic6be9cb3734b100a31465a0f2e75ea94861ffc10

Review URL: http://codereview.chromium.org/3438003