Add upstream git ebuild from portage repository.
This ebuild is dev-vcs/git in the upstream portage repository, and
depends on packages that have been moved, so we're going to have to
make some changes to it before it will work. Fortunately, it's masked
upstream, so committing this ebuild doesn't break anything.

TEST=emerge -1 git, make sure it doesn't use the new ebuild yet
BUG=chromium-os:6453

Change-Id: I3dbc73b44ece5b9828630c4b0cd6917f6316774c

Review URL: http://codereview.chromium.org/3426009