Fix saft / firmware collision.
When files were moved from the firmware package to the saft package, it
introduced a collision: The new version of saft collides with the old
version of the firmware package. To fix this, we need to ensure that
chromeos-firmware is upgraded before the new saft is installed.

TBR=vbendeb@chromium.org
BUG=chromium-os:7062
TEST=Emerge new saft when old chromeos-firmware is installed

Change-Id: I48a728d3d6d1e74d289b3ea7f0d1b0336aecfd09

Review URL: http://codereview.chromium.org/3397029