Make changes to the correct file
Change-Id: I61b3127d650678a1787e5cd6734e037eef5b9efd

The earlier submitted CL (http://codereview.chromium.org/3453019/show) was originally
changing both instances of vboot_reference ebuild, whereas
it was supposed to be changing just the -9999 version.

It was edited, but the wrong file was dropped from the cl.
This CL fixes the problem, restoring the changed ebuild and
changing the 9999 ebuild.

BUG=chromium-os:6614
TEST=manual

Review URL: http://codereview.chromium.org/3453023