This removes the default X root window cursor so that it doesn't show when we shut down and start up.
Someone else had this problem a long time ago, so this was basically
just adding a configure switch for compiling the xorg server.

This also maps all of the X cursors that we don't define to be the "left arrow" default pointer.  That way we don't get any surprises, and can easily see which ones map to that pointer.

BUG=7132
TEST=ran on device, tested multiple startup/shutdown scenarios and verified that X cursor was not shown.

Review URL: http://codereview.chromium.org/3522014

Change-Id: I8fedf5fb3d4dc60c6dfdedac53d7087a88242087