htpdate: Added sanity check against build time
Created a sanity check to make sure the time obtained from the remote
server is not before the build time.  It generates a timestamp that is
passed into htpdate as a #define.  htpdate's getHTTPdate function
compares it to the server time and throws an error if the server time is
before the build time.  That is an unrealistic scenario.

BUG=chromium-os:6905
TEST=verified failure case by passing in artificial build times that are
in the future

Change-Id: Id86b882e346763ef7344b8518237554e5cbdebaa
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/3535013