chromeos-initramfs: add dm-verity support
This change adds dmsetup+libs and rootdev to the initramfs.  In addition, it
adds device mapper detection and setup to the early userspace.

This is needed because the initramfs boot path avoids
calling prepare_namespace in the kernel.  Doing so skips
the device mapper boot-time setup -- so we do it here.

This change is dependent on a small build image change to ensure
dm-verity waits for the device to come up.

BUG=chromium-os:7452
TEST=built, imaged, and made it to the "wait for ethernet screen"

Review URL: http://codereview.chromium.org/3903003

Change-Id: I1219bdb2b8671599efeb8d8933975a4538afb2f7