Modify the light sensor startup script on a per-platform basis.
This works with a light-sensor.conf change that uses the variable and a platform overlay make.conf change that specifies the tuning value.

The net effect will be that different platforms will get different values in the TUNEVAL variable.

While this ebuild change and the light-sensor.conf script change and the make.conf change all work together,
they can be checked in in any order.  The ebuild will not fail if the expected string is not found.

BUG=chrome-os-partner:591
TEST=Manual.  Exposure to light and dark affect the backlight.

Change-Id: If3ef19a3beebe4ef2036d823e8ff26b35c8b4a5c

Review URL: http://codereview.chromium.org/4029004