build: Update chrome's strip flags.
The KEEP_CHROME_DEBUG_SYMBOLS variable can be set when
building chromeos-chrome to modify the strip command's flags
to retain some symbols in the binary.  The flags didn't
appear to be working when I tried to use this, though; the
binary appeared to be unsymbolized.

This change makes us use "--strip-debug --keep-file-symbols"
instead.  With these set, I'm seeing
base::debug::StackTrace::PrintBacktrace() symbolize its
output.

The variable is misnamed at this point since we're actually
stripping the debug symbols, but I think that
FEATURES=nostrip is the right way to retain them (and end up
with a Chrome binary that's way too big to actually put on a
device).

Change-Id: I830e0181f90bf3aeda7fc5f8d0cf5b830c9a7074

BUG=chromium-os:7470
TEST=build an image and tried it; image size increases by five megs or so and i get symbols

Review URL: http://codereview.chromium.org/4178004