htpdate: fix Date header parsing
A truncated or improper Date header may result in an out of bounds read.  This
may lead to an invalid date during parse or a sigsegv if it reads into a guard
or unallocated region.  At worst this is a crash or information leakage.

TEST=built for x86-generic
then ran it against www.google.com
ran it against
  echo "HTTP/1.1 200 OK
Date:
" | nc -l -p 9999

and again with good dates and truncated dates.  The parsing may still go crazily off base but it won't access data that was already in the buffer or past the buffer itself to other values on the stack.
BUG=chromium-os:8914

Change-Id: Iffb6d239cc91b5396ffd4978f32202594ed7a78a

Review URL: http://codereview.chromium.org/5331010