Added support for boards with only private overlays.
If a board has a public overlay it will be used as the primary overlay.
However, with this change, if a board only has a private overlay it will
be used as the primary.

BUG=7339

TEST=Ran setup_board, build_packages, and build_image on a board with only
a private overlay (locally created). Ran setup_board, build_packages, and
build_imge on a board with both public and private overlays.

Change-Id: Ib5de567ba1c62047969a20f3bea5e1db2bda020e

Related Issues:
http://codereview.chromium.org/3544009
http://codereview.chromium.org/3622003

Review URL: http://codereview.chromium.org/3571015