Allow extra board flags to be passed into build_packages.
This feature is useful for adding options to the emerge command in
build_packages. E.g. if you want to pass in the --emptytree option you
can just set the EXTRA_BOARD_FLAGS variable appropriately.

I'm planning to use this feature in cbuildbot.py so that we don't need
to delete the board every time we need to invalidate all the local
packages.

BUG=chromium-os:10127
TEST=EXTRA_BOARD_FLAGS='-ep' build_packages

Change-Id: Iaacc9e86baa1734dcb5acc6e285e6fd5413bbb4c

Review URL: http://codereview.chromium.org/5525003