Update host and target to retry on 500 errors.
wget treats 500 Server Error as a permanent error. It does not have a retry
feature for these types of errors. To get Portage to retry for 500 errors,
we need to switch to curl. This isn't in the stage3 currently, so we need to
wait until the system package has at least been emerged.

curl is already included in our system package, but I added it to the system
emerge line just to be safe.

BUG=chromium-os:10216
TEST=Test that curl retries on 500 server errors with the provided options.
     Run make_chroot --replace, setup_board and build_packages

Review URL: http://codereview.chromium.org/5680008

Change-Id: I94ffd74777204fa3e3d7ed3d58d6dddb40909f00