Switch to Mesa 7.9.
This update switches us to Mesa 7.9. This should help a little with
performance, and also with a number of OpenGL(WebGL) bugs we're seeing. This
also allows accelerated layers in chrome to mostly work (but it's still
disabled by default in session_manager_setup for now).

Change-Id: I7f2df310832131497c9e2e1bd41755b4c20fd0c8

BUG=
TEST=

Review URL: http://codereview.chromium.org/5925002