keyboard: Don't clear Control on Control-Alt-Backspace.
I think I was overzealous in clearing the modifier keys
here; we should be clearing Alt (as it's used to convert the
Backspace into Delete), but not Control.  This is needed for
the user to be able to type Ctrl-Alt-Shift-Backspace to
produce Ctrl-Shift-Delete to bring up the "Clear Browsing
Data" dialog.

I still need to make a theme change to delete GtkTextView's
hardcoded Ctrl-Shift-Delete "delete to the end of the
paragraph" binding; until then, this shortcut won't work
when the focus is in the omnibox.

(Yes, this is a difficult-to-type combination and hitting
Ctrl-Alt-Backspace still makes me cringe even though X isn't
configured to intercept it.  Seems worthwhile to stay
consistent with other platforms as long as we can, though.)

Change-Id: I68540fdc270df2a13e3487f3bdc8866d2d0a6fea

BUG=chromium-os:8986
TEST=built and tried it

Review URL: http://codereview.chromium.org/5958007