Simplify gpsd ebuild and use the latest available version.
This change modifies gpsd ebuild inherited from gentoo to
match the ChromeOS environment. The link is being renamed
from gpsd-2.38-r1.ebuild to gpsd-2.95-r2.ebuild to ensure
that the latest available version is downloaded.

All various GPS sensor protocols compiled in by default are
being disabled, the only protocol left in is "Generic NMEA"
which is what the Mario hardware complies with.

The ebuild src_unpack() function is modifying the setup.py
script to eliminate the run time python dependency.

The startup file gpsd.conf is modified to comply with the
upstart init daemon requirements and to wait for the GPS
device to be created and to enable the GPS before starting
gpsd.

Change-Id: Ie946c850ac7df14a92f6838e65aa7e05ec4119b4

BUG=chromium-os:467
TEST=manual

1. build and install the new image
2. reboot the ChromeOS device
3. being close to open sky run `cgps' from the command line
4. observe dynamic output showing current GPS satellite
   data and results of the fix. (the data should be flowing
   resulting in constant table updates, at least four
   satellites should be visible and the fix should be
   acquired within a few minutes or much less depending on
   previous GPS device activity.

Review URL: http://codereview.chromium.org/6025011