This should fix the binutils issue we were facing.
This change makes all symlinks relative. In the long run we probably
want all symlinks to be created in postinst().

Change-Id: Id097524cb72b19229d74e8acbf4a94744f6895c5

Redid the linking so it is relative instead of absolute.

Absolute linking gives problems when a binary toolchain is installed
using build_tc.py.
In fact, I believe linking should be done in postinst not in src_install
because otherwise the gold symlinks are broken.

Change-Id: I32470007cbc6ead7ca62f5eb975d7bf2ae1fe4af

BUG=11002
TEST=Emerged binutils using source. Then emerged using binary packages. No broken symlinks were found.

Review URL: http://codereview.chromium.org/6254005