Patch gtk+ ebuild to avoid installing unused i18n data under /usr/share/locale/.
This CL reduces the image size by more than 20MB. The ebuild patches the
configure file to remove support for optional languages. For some reason the
configure file is pregenerated and included in the gtk+ source tarball. The
patch is conditional on the "nls" USE flag which is unset globally in make.conf.

BUG=chromium-os:8066
TEST=x86-mario -- build_packages, build_image, tested on device,
used different Chrome languages and input methods.

Change-Id: Icaf446e780fb6e237d4141b217083ea9c85155c7

Review URL: http://codereview.chromium.org/6246014