Move the creation of the firmware layout configuration file to the chromeos-bios ebuild.
This also fixes emerging chromeos-u-boot-next for the generic tegra2 overlay.

This CL is accompanied by CL: http://codereview.chromium.org/6309024/ which removes this functionality from U-Boot.

The firmware_layout_config file is a copy of the file that is removed by the aforementioned CL.  It comes from the top level of the u-boot-next.git repository.

Change-Id: I10743dda9d782874da997de5ebb423685c28e30a

BUG=None
TEST=emerge-tegra2_seaboard chromeos-bios; verify that image.bin is created.  emerge-tegra2 chromeos-u-boot-next; note that it succeeds now.

Review URL: http://codereview.chromium.org/6332021