Remove setting up the /usr/bin/X11/X symlink from customize_rootfs.
This symlink is setup in the xorg-server ebuild now. A follow on cleanup change
could remove the symlink if necessary -- but it needs to update various clients
such as login manager and several autotests to point to /usr/bin/X rather than
/usr/bin/X11/X.

BUG=chromium-os:11540
TEST=x86-mario -- build_packages, build_image, booted the device, made sure the
symlink is still there.

Change-Id: I3367525d799a32a2293742854d3229f0709b2995

Review URL: http://codereview.chromium.org/6312071