Verify that SCRIPTS_DIR exists in build_packages and setup_board.
This hack is necessary for old chroots to ensure that we source a good
common.sh when the one inside the chroot is out of date. We'll be able
to get rid of this after everybody has updated their chroot to have a working
version of common.sh and/or we have deprecated running scripts from the
source directory.

Also fix logic for when common.sh is missing to exit properly.

BUG=chromium-os:11598
TEST=Verify that build_packages and setup_board still work when
     old common.sh is present in chroot.

Change-Id: I88431decb50ee17c239d081fa0cc6873685b19e4

Review URL: http://codereview.chromium.org/6334076