Default to building all U-Boot configs.
This is required in order to start using the chromeos-bios ebuild.
Portage can't manage depending on a virtual with use flags
specified.  So we can't depend on the u-boot virtual and specify
all of these use flags.

Change-Id: I86c7d90e1b96842d94fe411786c60c617f5d1243

BUG=None
TEST=emerge-tegra2_seaboard -av chromeos-u-boot-next

verify that all configs are built (including the legacy u-boot.bin).

Review URL: http://codereview.chromium.org/6286123