Remove device preseeding from customize_rootfs.
Based on the comment, it seems that this was done temporarily for debugging
purposes. chromeos_startup does the copy. If the copy is really needed as part
of the build, it can be done in the chromeos-init ebuild.

BUG=chromium-os:11592
TEST=build_image, booted

Change-Id: Icd20e864f4d28d6dcfc6de0c87951c726c72551d

Committed: http://chrome-svn/viewvc/chromeos?view=rev&revision=d4e786b

Review URL: http://codereview.chromium.org/6246086