Update laptop-mode-tools to behave the same with new Portage.
The new version of Portage checks that all use flags are declared, so it
doesn't allow us to forget to declare them anymore. Declaring them has
the side effect of enabling the ACPI flag (which is explicitly enabled
elsewhere), so I disable ACPI to match the previous (maybe broken) behavior.
This fixes the build. In a follow-up commit, we'll address what flags we
want to enable or not.

TBR=snanda@chromium.org
BUG=none
TEST=Check that laptop

Review URL: http://codereview.chromium.org/6475017

Change-Id: I8f62791332b8b69b2c134bc5595d663d0074db03