Remove incorrect PROVIDES define in chromeos-u-boot-next-build-env.
This was causing a problem because it was fooling portage into
thinking that U-Boot had built correctly once this ebuild
built and installed.  But this ebuild does not actually
provide U-Boot, it just provides a copy of the U-Boot build
environment for the V-Boot reference code to use.

BUG=None
TEST=None

Change-Id: I28b3c03dcaccca0149d9a0e784b31929a149e090

Review URL: http://codereview.chromium.org/6546059