Setting up virtual/chromeos-bsp with a null implementation
Folks - this is an attempt to replace the intended functionality of board-devices.ebuild with a new name, and using the idea of "virtuals" instead of overlaying on the board-devices-private name.

Once this is landed, then individual board overlays (whether they are from the board-overlays.git or other repositories) can redefine the meaning of virtual/chromeos-bsp to some ebuild that they provide.

BUG=None
TEST=Manually emerged virtual/chromeos-bsp

Review URL: http://codereview.chromium.org/6507036

Change-Id: I0021fa4ea4de426854e45413dd9fa40b3ebe547f