vpd: only depends on flashrom for x86 platform.
The flashrom is not verified on ARM platform yet and actually the VPD utility is
not necessary to depend on flashrom.

Hence, remove the dependency for ARM so that vpd can be compiled on ARM. Will
add flashrom back once it is verified.

Also fixed the bug of libuuid. It was accidentally added in RDEPEND. It should be in DEPEND.

Change-Id: If3a317ec228ddb50c4f1555db3ff868ca2bb1c64

BUG=chrome-os-partner:1496
TEST=run the following commands
emerge-x86-generic vpd
emerge-tegra2_dev-board vpd

Review URL: http://codereview.chromium.org/6541067