Update host and target to retry on 500 errors.
wget treats 500 Server Error as a permanent error. It does not have a retry
feature for these types of errors. To get Portage to retry for 500 errors,
we need to switch to a new version of curl. This isn't in the stage3 currently,
so we need to wait until the system package has at least been emerged.

curl is only included in our system package due to the curl
use flag for git, so we add it to the emerge line to be sure
it is installed.

BUG=chromium-os:10216
TEST=Test that curl retries on 500 server errors with the provided options.
     Test that curl fails on 404 errors with the provided options.
     Run make_chroot --replace, setup_board and build_packages.
     Run emerge -1 iotools (the case that failed last time we committed)

Review URL: http://codereview.chromium.org/5680008

Change-Id: I47c6f2e65789c4ec6a6ad93754ef82a079962606