Remove the rm -rf of /etc/ssh from the openssh ebuild.
Also uprev the openssh and ssh-known-hosts ebuilds to force
all chroots to reinstall these ebuilds (or just the
ssh-known-hosts ebuild if that's all the chroot has left in it)
and add a dependency from ssh-known-hosts to openssh.  The
dependency forces the new version of openssh to be installed
(and thus the old version to be removed, which does an rm -rf)
before the ssh-known-hosts ebuild installs it's file in
/etc/ssh.

BUG=chromium-os:12484
TEST=sudo emerge ssh-known-hosts
Verify that openssh is upgraded and /etc/ssh still exists.

Change-Id: Idbbd15ba624ce787d8b2af3457f73e38ca7a5daf

Review URL: http://codereview.chromium.org/6581052