Use new flash.cfg to build better BCTs.
The flash.cfg file installed by the tegra-bct-<board> ebuilds describes
the parameters for the boot flash device.  These values are added to the
cbootimage config file so that cbootimage will modify the resulting BCT
to match the boot flash device.  If the flash.cfg file is empty then
the BCT file provided to cbootimage will not have it's flash parameters
modified and will be used as is.

Change-Id: I080d1df026d2236b809133749f3eb42a558ca842

BUG=chromium-os:11981
TEST=use as yet to be checked in cros_write_firmware to write and boot a board from SPI.

This CL depends on: http://codereview.chromium.org/6578027/

Review URL: http://codereview.chromium.org/6581042