Rename customize_rootfs to cros_set_lsb_release because this is all it does.
We should probably move cros_set_lsb_release into an ebuild but that project
will require quite a few changes and it's not completely clear it's needed. For
now just rename customize_rootfs to signify what it does and to discourage
further additions to that script.

customize_rootfs is a symlink now because build_image lives in a separate repo
-- to be updated in a separate CL.

BUG=chromium-os:11647
TEST=build_image, checked /etc/lsb-release

Change-Id: I87cd15a271c6acb96d5723d6a450611dd7134b7b

Review URL: http://codereview.chromium.org/6598080