Update kernel to accept use flags based on kernel config.
This change helps differentiate between different kernel configs so that
we don't reuse the same binaries across different boards even if we're
reusing the same binary on different boards.

Once this change is in and the overlays are defining the appropriate use flags,
I am planning on getting rid of the kernel config environment variables and
just using use flags to select the configuration.

Change-Id: Ic47e81706007e296ec0bb2bdd0a10392bc648f82

BUG=chromium-os:11743, chromium-os:10962
TEST=Verify that the use flags are available and can be set in the overlay config.

Review URL: http://codereview.chromium.org/6628004