Support platform specific light sensor settings.
This change is backward compatible so the script name can be changed
without breaking a build.

For this to actually do anything, power_manager and kernel changes need to
also be committed.  But this one is best to commit first.

Change-Id: I14b6b2593c80035d4614b836cd3778b849283392

Signed-off-by: Bryan Freed <bfreed@chromium.org>
Signed-off-by: Benson Leung <bleung@chromium.org>

BUG=chrome-os-partner:2163
TEST=manual: see if /sys/class/iio/devic0/ is populated.

Review URL: http://codereview.chromium.org/6632008
Patch from Bryan Freed <bfreed@chromium.org>.