Update perf to new-style virtual package
virtual/perf was apparently an old-style virtual package because of the
need to select dev-util/perf on x86 and dev-util/perf-next on arm.
However this then did not allow a board's profile to change the virtual
package, ie. kernel-next ended up building the wrong version of perf for
the kernel.

BUG=chromium-os:12897
TEST=ran emerge -p virtual/perf for x86-generic, x86-mario w/profile & arm-generic

Change-Id: I678a50c9d1dd8b5eca5d8ccca73e6b27f59b545b

Review URL: http://codereview.chromium.org/6652005