MT: Initial ebuilds for MT input support.
These are ebuilds for the open source mtdev and xf86-input-multitouch packages, in addition to an xorg conf change.

To use the new software, set the 'multitouch' USE flag.

I won't check these in until I've tested them and checked in the source to the git repos.

BUG=chromium-os:12967
TEST=Will compile/run before checking in

Review URL: http://codereview.chromium.org/6660018

Change-Id: I7386f4d688e6c76050620eade8b84746fba2dfb5