Apply Brian's LCD sharpening patch to Freetype 2.4.4
Freetype was updated to 2.4.4 to pull in a security patch (Tretype GX fonts) and bug fixes for CFF outline handling. After that, we have to reapply Brian's LCD sharpening patch.

BUG=chromium-os:11982
TEST='emerge-<foo> freetype' builds Freetype 2.4.4-r1. 'gmerge freetype'
and restart ChromeOS. The change is subtle but careful inspection of the
glyphs should show them to be slightly slimmer and sharper.

Change-Id: Ifc3f5e472b54d49ed81b5d4e3c88092f7a41b5b3

Review URL: http://codereview.chromium.org/6688005