Update CrOS core fonts
The following fonts are updated to 1.20
  - Arimo : Regular, Bold, Italic, Bold-Italic
  - Tinos : Regular, Bold, Italic, Bold-Italic
  - Cousine : Regular, Bold, Italic, Bold-Italic

The most significant change in 1.20 is that Hebrew is now covered with
Arimo, Tinos and Cousine in all variant styles (r,b,i, bi) above.

Other changes include:

- Tuned the hints on all glyphs for byte-code interpretation on Chrome OS.
- Tested for hinted metric compatibility with the Arial/TNR/Courier New  fonts up to 128 ppem.
- Added GPOS kerning for the Latin/Greek/Cyrillic which matches the data in
  their respective 'kern' tables.
- Fixed the shape of the polytonic Omega glyphs in the Bold Italic.
- Fixed T-cedilla and T-comma glyphs
- Fixed the glyph for U+04BE (Cousine)

In a follow-up CL, a stand-alone Hebrew font (Droid Sans Hebrew) will
be removed. The Hebrew glyphs in Arimo are identical to a new version of
Droid Sans Hebrew (which we don't include) and are rather similar
to the old version of Droid Sans Hebrew that will be removed.

Symbol Neu is not changed and is still 1.0.

The ebuild version was bumped from 0.0.5 to 1.20.0 to match the font version (inside TTF name table).

BUG=chromium-os:13423
TEST='emerge croscorefonts' works fine. No apparent change in the UI.

Change-Id: Icd4a9c7e40828df7c516d920e9fbdda885924c84

Review URL: http://codereview.chromium.org/6676095