[chromiumos-overlay] Make protobuf ebuild cross-compile friendly, and uprev
When cross-compiling python support in protobuf, we need to do two
things to have the package build cleanly:

1) pass --with-protoc=/path/to/host/protoc
2) convince the python package build system (python/setup.py) to use the host
   protoc as well.  This requires a patch to the upstream source and setting
   an env var in the ebuild.

Move up to a patched version of protobuf-2.3.0 that enables this, and
mod the ebuild to enable cross-compile.
We also mod the ebuild to add a USE flag that requires python to be
available at build time only.  This should be removed when
http://code.google.com/p/chromium-os/issues/detail?id=13429 is fixed.

BUG=13282
TEST=build_packages for arm- and x86-generic

Change-Id: Ib3b6ebd0e3256c422cb12714bbcb9f15bdf40c2b

R=davidjames@chromium.org

Review URL: http://codereview.chromium.org/6729007