Create chromeos-flasher-script ebuild and flasher scripts.
The spi.script file comes from the cros-devutils repository.  Initially
it was installed with the cros_write_firmware tool.  But installing
it and the nand.script file as part of the board build makes more
sense.  This way the bootflash-* USE flag is used to select the
device that U-Boot will be written to.

BUG=chromium-os:11981
TEST=emerge chromeos-flasher-script for Seaboard and verify that the
spi.script file is installed.  Emerge for Harmony and verify that the
nand.script file is installed.

Change-Id: Id04f01480de77fa0078f4c3318c311ec34c0db3d

Review URL: http://codereview.chromium.org/6730035