devicekit-power : Support batteries that report charge and energy.
Make dkp-device-supply more clean in the way it treats
charge and energy. Previously, if energy was present,
it would assume that current was actually energy_rate.

Now, if both are present, it will treat current as current in UA
and energy as energy as uWh.

Change-Id: I98c70fc254a20a07cdb24a33135095d8e38ce4bc
Signed-off-by: Benson Leung <bleung@chromium.org>
Signed-off-by: Sameer Nanda <snanda@chromium.org>
Signed-off-by: Rhyland Klein <rklein@nvidia.com>

R=snanda@chromium.org
BUG=chrome-os-partner:1781
TEST=Try this patched devicekit-power on a system that reports
both current and energy. Ensure correct units are reported by
devkit-power -d for energy_rate.
Try the same on a system that only reports current. Ensure
energy_rate is correct.

Review URL: http://codereview.chromium.org/6713079