[chromiumos-overlay] Make protobuf ebuild cross-compile friendly, and uprev
When cross-compiling python support in protobuf, we need to do two
things to have the package build cleanly:

1) pass --with-protoc=/path/to/host/protoc
2) convince the python package build system (python/setup.py) to use the host
protoc as well. This requires a patch to the upstream source and setting
an env var in the ebuild.

Move up to a patched version of protobuf-2.3.0 that enables this, and
mod the ebuild to enable cross-compile.
- We also modify the ebuild to add a USE flag that requires python to be
available at run time.
- DEPEND on python.
- Install python libs to /usr/local so that they live alongside the dev packages.
These should be removed when
http://code.google.com/p/chromium-os/issues/detail?id=13429 is fixed.

BUG=13282
TEST=verify source package requires python, but binary package does not

Change-Id: I843a4757408cd4232a98b1f2352cd924b0865439

R=davidjames@chromium.org

Review URL: http://codereview.chromium.org/6670093