sys-apps/mosys: make ebuild behave like a real Gentoo ebuild.
The LDFLAGS "workaround" shouldn't have been present in the first place;
export the tools just the same both on cross-compile and not; then export
LDFLAGS to be "raw" (i.e. without the '-Wl,' prefix), since it uses ld(1)
directly and not the gcc frontend.

Change-Id: I5409951c9d4cbef26dca7522e704fecebdbedf01

R=dhendrix@chromium.org
BUG=none
TEST=built straight, x86-generic and arm-generic

Review URL: http://codereview.chromium.org/6756028