Remove redundant steps in src_test() for chromeos-wm ebuild
All unit tests use the automatic portage flow for determining which
ebuild steps are required, such as when compile is required.  So it's
ok to remove the workaround redundant steps in src_test() of ebuild.

Note that this change must not be pushed until after http://codereview.chromium.org/6812021.

BUG=chromium-os:13634
TEST=./cros_run_unit_tests --board=x86-mario --packages=chromeos-wm

Review URL: http://codereview.chromium.org/6812008

Change-Id: Id82b2975b11b6d9cbec18a1ee5835257daafe5d4