app-misc/utouch-evemu: add SIGINT and SIGTERM to exit
A signal handler is added to handle SIGINT and SIGTERM. The output file is
closed properly before the program exits. This is useful in trackpad record
program to terminate evemu-record earlier by issuing either of the above said
signals.

BUG=chromium-os:14096
TEST=On a chromuim os machine:
(1) $ sudo evemu-record /dev/input/event6 > /tmp/packets.out
(2) Use your finger to move on the trackpad. This is supposed to generate
trackpad packets to /tmp/packets.out
(3) Issue SIGINT by pressing Ctrl-c;
or issue SIGTERM by typing '$ kill -15 pid'
(4) Type '$ cat /tmp/packets.out', and observe the last line to see whether it
is truncated or not. In this patch, a signal handler is added for capturing
SIGINT and SIGTERM, and close the file before exit. Hence, the last line should
not be truncated. The last line should look something like the following line
with the last three numbers as "0000 0000 0".

E: 1304572397.518946 0000 0000 0

Change-Id: I6ffe2bbe44a23d5b0cd02307e086d0e741132ae6
Reviewed-on: http://gerrit.chromium.org/gerrit/325
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Daniel Kurtz <djkurtz@chromium.org>