Always build recovery and developer firmware images as well as normal.
Always building the recovery image is required so that the images
that come off of the build bot will have firmware that can be used
to boot existing systems.  It seems like providing both of these
images in the build is useful and simplifies the factory build
process slightly.

BUG=None
TEST=Build image for Kaen and run clobber_firmware.  Verify the result boots.

Change-Id: Ieb721d401a00176486f0da0562c2c7651fee274d

Review URL: http://codereview.chromium.org/6898030