Modified gcc-9999 ebuild to allow building gcc-4.6.0.
gcc-9999 now has a use flag to build upstream compilers.

The binary and data paths of gcc install have changed to allow for
side-by-side installation of 4.4.3 and 4.6.0.

A glibc patch was added to allow building with gcc-4.6.0. gcc-4.6.0
has a default command line flag that needs to be reversed in the
Makefile.

gcc-4.6.0's ebuild has been added. emerge that with USE=multislot to
emerge gcc-4.6.0 side-by-side with gcc-4.4.3. Switching between
compilers can be done by using gcc-config.

setup_board was changed to accomodate changes in gcc paths. The location
of libgcc_s.so was hardcoded earlier. equery is now used to determine
the correct location of the libgcc_s.so file.

Fixed some formatting issues with the 9999 ebuild.

Copied the 9999 into 4.6.0 after fixing the formatting issues.

Addressed issued pointed to by Raymes:
1. Added ccache linking and removal code.
2. Made it so that GCC_PV is obtained from the ebuild version.
3. Copied hardened specs into a separate file.
4. Miscelleneous other changes.

Rebased gcc-9999.ebuild.

Added a better way of detecting whether the toolchain is installed in
setup_board.

Uprevved the glibc ebuild to r3. Changed setup_board to use that instead
of r2.

Change-Id: Ib6f0a9a244f3d3fa670a5ab071a93e408a5bf810