Add clang to hard-host-depends.
We want clang in the chroot so that we can use its static analysis.
We're not currently planning to use it for code generation.  The
union of the warnings from gcc and the warnings from clang should allow
us to catch more bugs than the warnings from just gcc or just clang.

This change does not cause anything to use clang; it just adds clang
to the chroot.

BUG=none
TEST=make chroot, enter it, run clang --version

Change-Id: Ie0792b2093d4f351a5d2ee8d23618ea229665a49
Reviewed-on: http://gerrit.chromium.org/gerrit/710
Reviewed-by: Anush Elangovan <anush@chromium.org>
Tested-by: Geoff Pike <gpike@chromium.org>