laptop-mode-tools: Skip failed globs when finding module scripts
Globs return the glob expression itself if there are no valid expansions  (Unless using the bash-specific "setopt -u nullglob").
Fix this by detecting and ignoring SCRIPT when it is not a file.

BUG=chromium-os:15187
TEST=(0) Enable laptop_mode verbose output (set VERBOSE_OUTPUT=1 in /etc/laptop-mode/laptop-mode.conf)
(1) sudo /usr/sbin/laptop_mode auto force
The following should not be present at end of output:
Module /usr/local/lib/laptop-mode-tools/modules/* is not executable or is to be skipped.
Module /usr/local/share/laptop-mode-tools/modules/* is not executable or is to be skipped.
Module /etc/laptop-mode/modules/* is not executable or is to be skipped.

Change-Id: Ic125b132476ce33aa20d83e780c056f216cf1fff
Reviewed-on: http://gerrit.chromium.org/gerrit/671
Reviewed-by: Simon Que <sque@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>