Adjust fontconfig for Chinese and Korean fonts.
This is done for two reasons:

a. To get Chinese text look better when the UI langauge is not Japanese
nor Korean. That is, it'll make Chinese text look better in Chinese
locales as well.
b. To get Korean Hangul look better when the UI language is not Korean.

This CL does the following:

1. Do not use embedded bitmap glyphs in PMingLiU/MingLiU/Song ASC/N Song
ASC.
2. Turn on light autohints for Song ASC/N Song ASC and BCI hints for
(P)MingLiu. PMingLiU and MingLiU have component glyphs that are moved
and scaled with byte code instructions and autohinting does not
work well.
3. Turn off sub-pixel anti-aliasing for the above fonts and Droid Sans
Fallback.
4. Put Song ASC/N Song ASC, PMingLiU/MingLiU and NanumGothic before Droid Sans
Fallback. (see http:/crosbug.com/14691). Note that PMingLiU and Song ASC
are serif fonts so they're not supposed to be in sans-serif section, but
until bug 14691 is resolved, we have to put them before Droid Sans
Fallback.
5. Remove the unnecessary orthography patch.

BUG=chromium-os:8757
TEST=Go to http://news.google.hk and characters look 'prettier' than
before overall.

Change-Id: Ib22838d13e68245f5716d93a97b305e072bd2b35
Reviewed-on: http://gerrit.chromium.org/gerrit/709
Reviewed-by: Brian Stell <bstell@chromium.org>
Tested-by: Jungshik Shin (jungshik at google) <jshin@chromium.org>