Add git/repo branch info to PS1 for new chroots.
The format looks like this:
  (cr) ((branch/f9fa9f5...)) user@machine ~/trunk/src/scripts $

NOTE that old chroots will not be fixed up.  Since this is just an
optional fix (and also a bit of a matter of preference), we're just going
to modify the default.  If someone really wants the new PS1, they can
manually modify their ~/.bashrc to have this for PS1:
  export PS1='$(__git_branch_ps1 "(%s)")'" ${PS1}"
  export PS1="(cros-chroot) ${PS1}"

ALSO NOTE that if someone really hates the new PS1, it's not a big deal
for them to go into their .bashrc and change it.  We won't clobber it.

BUG=chromium-os:13186
TEST=Manually ran the new bit of code in make_chroot and saw the proper
bits generated in my ~/.bashrc.  Then, exited/entered the chroot and saw
the appropriate prompt.  Changed to a different git project and saw the
prompt change.  Manually removed the hiding of 'master' and saw that
the branch name (master) now showed up for me.

Change-Id: If5a7a434ae76b5449310dfc5f5df5a082f8e84aa
Reviewed-on: http://gerrit.chromium.org/gerrit/1584
Tested-by: Doug Anderson <dianders@chromium.org>
Reviewed-by: Darin Petkov <petkov@chromium.org>
Reviewed-by: Puneet Kumar <puneetster@chromium.org>