Don't ignore uprevved ebuilds with --rebuilt-binaries.
We want to update our buildbots to use the rebuilt-binaries flag to
track a binhost. Unfortunately, there is a bug with the flag that is
preventing us from doing that.

When --rebuilt-binaries is used, packages from the binhost are
preferred over unbuilt packages, even if the package has been uprevved.
That would cause our preflight buildbot to not include changes from
developers. This patch for Portage from upstream fixes the above bug.

BUG=chromium-os:15810
TEST=Verify above bug is fixed with new version of Portage.

Change-Id: I77d558741df825e8b3a3d7600b995bbbbfd9ae9b
Reviewed-on: http://gerrit.chromium.org/gerrit/1635
Reviewed-by: Anush Elangovan <anush@chromium.org>
Tested-by: David James <davidjames@chromium.org>