upstart: log to kmsg when built USE=upstartdebug
One of the problems with debugging Upstart is that its default log
priority is low, and it sends to syslog which doesn't start until late.

For maximum debugging you want a log priority of INFO (equvialent to
--verbose on the kernel command-line, which we can't easily change while
testing) and you want the messages sent to the kmsg ring-buffer so you
can retrieve them from boot.

This isn't on by default because there's a slight perfomance penalty to
write these messages compared to syslog, so it's behind a USE flag; just
USE=upstartdebug emerge-${BOARD} upstart
or USE=upstartdebug gmerge upstart
to turn it on.

BUG=chromiumos:15876
TEST=built with and without USE and compared

Change-Id: Ia3b2f77aa10c0cbf4cbff63a64e59674aaadc919
Reviewed-on: http://gerrit.chromium.org/gerrit/1738
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>
Tested-by: Scott James Remnant <keybuk@chromium.org>